Fix Dirwalk bug when compiled in x64 mode
Good night! Have a nice dream!

+++ dirop_win.cpp
@@ -7,7 +7,7 @@ struct DirWalk::_impl
 {
 	int _status;
 	string _dirname;
-	int _handle;
+	intptr_t _handle;
 };
 
 DirWalk::DirWalk() : _p(new _impl)
